Refactor geo importer
- Split geonames-sql.js into geonames.js and geotags.js.

- Renamed the schools and geonames tasks to:
  import:schools
  import:geonames
  import:geotags

- Used a better upsert method, allowing to run the tasks
  on a remote machine.

- Reduced memory consumption.
