+++ package.json
@@ -16,8 +16,9 @@
     "start": "NODE_ENV=development babel-node --source-maps -- index.js",
     "docker-start": "knex migrate:latest && NODE_ENV=production npm run webpack:build:prod && NODE_ENV=production babel-node index.js",
     "lint": "eslint src",
-    "geonames": "babel-node bin/geonames-sql.js",
-    "schools": "babel-node bin/schools.js",
+    "import:geonames": "babel-node bin/geonames.js",
+    "import:geotags": "babel-node bin/geotags.js",
+    "import:schools": "babel-node bin/schools.js",
     "webpack:build:dev": "webpack --config webpack.dev.config.js",
     "webpack:build:prod": "NODE_ENV=production webpack --config webpack.prod.config.js",
     "db:migrate": "knex migrate:latest",
